feat: table formatting refactoring
