+++ youmubot-osu/src/discord/announcer.rs
@@ -1,16 +1,5 @@
-use super::db::{OsuSavedUsers, OsuUser};
-use super::{calculate_weighted_map_length, OsuClient};
-use super::{embeds::score_embed, BeatmapWithMode};
-use crate::{
-    discord::beatmap_cache::BeatmapMetaCache,
-    discord::cache::save_beatmap,
-    discord::oppai_cache::{BeatmapCache, BeatmapContent},
-    models::{Mode, Score, User, UserEventRank},
-    request::UserID,
-    Client as Osu,
-};
-use announcer::MemberToChannels;
-use serenity::builder::CreateMessage;
+use std::{convert::TryInto, sync::Arc};
+
 use serenity::{
     http::CacheHttp,
     model::{
@@ -18,10 +7,25 @@ use serenity::{
         id::{ChannelId, UserId},
     },
 };
-use std::{convert::TryInto, sync::Arc};
-use youmubot_prelude::announcer::CacheAndHttp;
-use youmubot_prelude::stream::{FuturesUnordered, TryStreamExt};
+use serenity::builder::CreateMessage;
+
+use announcer::MemberToChannels;
 use youmubot_prelude::*;
+use youmubot_prelude::announcer::CacheAndHttp;
+use youmubot_prelude::stream::TryStreamExt;
+
+use crate::{
+    Client as Osu,
+    discord::beatmap_cache::BeatmapMetaCache,
+    discord::cache::save_beatmap,
+    discord::oppai_cache::{BeatmapCache, BeatmapContent},
+    models::{Mode, Score, User, UserEventRank},
+    request::UserID,
+};
+
+use super::{calculate_weighted_map_length, OsuClient};
+use super::{BeatmapWithMode, embeds::score_embed};
+use super::db::{OsuSavedUsers, OsuUser};
 
 /// osu! announcer's unique announcer key.
 pub const ANNOUNCER_KEY: &str = "osu";
